requirements.txt
# VPet Requirements
# Core dependencies
pillow>=10.0.0
pystray>=0.19.0
pywin32>=306

# AI Chat functionality (optional)
google-generativeai>=0.3.0
ollama>=0.1.0  # Added for local Ollama support

# Build tools
pyinstaller>=5.0.0

# Additional dependencies that might be needed
requests>=2.25.0
